简化Window AppTitleBar设置
