+++ src/Pixeval/Controls/IEnhancedPage.cs
@@ -1,8 +1,8 @@
-﻿#region Copyright (c) Pixeval/Pixeval
+#region Copyright (c) Pixeval/Pixeval
 // GPL v3 License
 // 
 // Pixeval/Pixeval
-// Copyright (c) 2023 Pixeval/RefreshDragRegionMessage.cs
+// Copyright (c) 2023 Pixeval/IEnhancedPage.cs
 // 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -18,9 +18,17 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #endregion
 
-namespace Pixeval.Messages;
+using Microsoft.UI.Xaml.Navigation;
 
-public record RefreshDragRegionMessage
+namespace Pixeval.Controls;
+
+public interface IEnhancedPage
 {
-    public static readonly RefreshDragRegionMessage Shared = new();
+    public int ActivationCount { get; }
+
+    public bool ClearCacheAfterNavigation { get; set; }
+
+    public bool Initialized { get; }
+
+    void OnPageDeactivated(NavigatingCancelEventArgs e);
 }
