+++ src/Pixeval/Pages/MainPage.xaml.cs
@@ -49,18 +49,17 @@ using Pixeval.Messages;
 using Pixeval.Pages.Capability;
 using Pixeval.Pages.Download;
 using Pixeval.Pages.Misc;
+using Pixeval.UserControls.IllustrationView;
 using Pixeval.Util;
 using Pixeval.Util.Threading;
 using Pixeval.Util.UI;
 using Pixeval.Utilities;
 using Image = SixLabors.ImageSharp.Image;
 using WinUI3Utilities;
-using Windows.Graphics;
-using IllustrationViewModel = Pixeval.UserControls.IllustrationView.IllustrationViewModel;
 
 namespace Pixeval.Pages;
 
-public sealed partial class MainPage : ISupportCustomTitleBarDragRegion
+public sealed partial class MainPage
 {
     private static UIElement? _connectedAnimationTarget;
 
@@ -86,7 +85,7 @@ public sealed partial class MainPage : ISupportCustomTitleBarDragRegion
     {
         // dirty trick, the order of the menu items is the same as the order of the fields in MainPageTabItem
         // since enums are basically integers, we just need a cast to transform it to the correct offset.
-        ((NavigationViewItem) MainPageRootNavigationView.MenuItems[(int) App.AppViewModel.AppSetting.DefaultSelectedTabItem]).IsSelected = true;
+        ((NavigationViewItem)MainPageRootNavigationView.MenuItems[(int)App.AppViewModel.AppSetting.DefaultSelectedTabItem]).IsSelected = true;
 
         // The application is invoked by a protocol, call the corresponding protocol handler.
         if (App.AppViewModel.ConsumeProtocolActivation())
@@ -146,7 +145,7 @@ public sealed partial class MainPage : ISupportCustomTitleBarDragRegion
 
     private async void KeywordAutoSuggestBox_GotFocus(object sender, RoutedEventArgs e)
     {
-        var suggestBox = (AutoSuggestBox) sender;
+        var suggestBox = (AutoSuggestBox)sender;
         suggestBox.IsSuggestionListOpen = true;
         await _viewModel.SuggestionProvider.UpdateAsync(suggestBox.Text);
     }
@@ -202,7 +201,7 @@ public sealed partial class MainPage : ISupportCustomTitleBarDragRegion
         using (var scope = App.AppViewModel.AppServicesScope)
         {
             var manager = scope.ServiceProvider.GetRequiredService<SearchHistoryPersistentManager>();
-            if (manager.Count == 0 || manager.Select(count: 1).AsList() is [ { Value: var last }, ..] && last != text)
+            if (manager.Count == 0 || manager.Select(count: 1).AsList() is [{ Value: var last }, ..] && last != text)
             {
                 manager.Insert(new SearchHistoryEntry
                 {
@@ -238,7 +237,7 @@ public sealed partial class MainPage : ISupportCustomTitleBarDragRegion
         await MainPageRootFrame.AwaitPageTransitionAsync<SettingsPage>();
         var settingsPage = MainPageRootFrame.FindDescendant<SettingsPage>()!;
         var position = settingsPage.FindChild<FrameworkElement>(element => element.Tag is SettingEntry e && e == entry)
-            ?.TransformToVisual((UIElement) settingsPage.SettingsPageScrollViewer.Content)
+            ?.TransformToVisual((UIElement)settingsPage.SettingsPageScrollViewer.Content)
             .TransformPoint(new Point(0, 0));
         settingsPage.SettingsPageScrollViewer.ChangeView(null, position?.Y, null, false);
     }
@@ -295,26 +294,6 @@ public sealed partial class MainPage : ISupportCustomTitleBarDragRegion
             .Build(CurrentContext.Window);
         content.Owner = dialog;
         await dialog.ShowAsync();
-    }
-
-
-    public Task<RectInt32[]> SetTitleBarDragRegionAsync(FrameworkElement? titleBar, ColumnDefinition[]? dragRegions)
-    {
-        var (leftDragRegion, leftMarginRegion, searchBarRegion, marginRegion, reverseSearchButtonRegion, searchSettingButtonRegion, rightDragRegion) = (dragRegions![0], dragRegions[1], dragRegions[2], dragRegions[3], dragRegions[4], dragRegions[5], dragRegions[6]);
-        const int leftButtonWidth = 50;
-        var scaleAdjustment = UIHelper.GetScaleAdjustment();
-        RectInt32 dragRectL;
-        dragRectL.X = (int) (leftButtonWidth * scaleAdjustment);
-        dragRectL.Y = 0;
-        dragRectL.Width = (int) ((leftDragRegion.ActualWidth + leftMarginRegion.ActualWidth - leftButtonWidth) * scaleAdjustment);
-        dragRectL.Height = (int) ((titleBar?.ActualHeight ?? 45) * scaleAdjustment);
-        RectInt32 dragRectR;
-        dragRectR.X = (int) (leftButtonWidth + (leftDragRegion.ActualWidth + leftMarginRegion.ActualWidth + searchBarRegion.ActualWidth + marginRegion.ActualWidth + reverseSearchButtonRegion.ActualWidth + searchSettingButtonRegion.ActualWidth) * scaleAdjustment);
-        dragRectR.Y = 0;
-        dragRectR.Width = (int) (rightDragRegion.ActualWidth * scaleAdjustment);
-        dragRectR.Height = (int) ((titleBar?.ActualHeight ?? 45) * scaleAdjustment);
-
-        return Task.FromResult(new[] { dragRectL, dragRectR });
     }
 
     private void KeywordAutoSuggestBox_OnDragOver(object sender, DragEventArgs e)
